==> data_module/1.0-Data_Collection.ipynb <==
<div style="text-align: center; background-color: #b1d1ff; font-family: 'Trebuchet MS', Arial, sans-serif; color: white; padding: 20px; font-size: 40px; font-weight: bold; border-radius: 0 0 0 0; box-shadow: 0px 6px 8px rgba(0, 0, 0, 0.2);">
Data Collection
</div>

## Import

import requests
import numpy as np
import pandas as pd
import re
import xml.etree.ElementTree as ET

## Collect data from website

def collect_data(cur_page, items_per_page):
    # Khai báo sẵn các mảng để đưa vào dataframe
    name = []
    region = []
    country = []
    overall = []
    AcademicReputation = []
    EmployerReputation = []
    FacultyStudentRatio = []
    CitationsPerFaculty = []
    InternationalFacultyRatio = []
    InternationalStudentsRatio = []
    InternationalResearchNetwork = []
    EmploymentOutcomes = []
    Sustainability = []
    
    url = "https://www.topuniversities.com/rankings/endpoint"

    querystring = {"nid":"3897789","page":f"{cur_page}","items_per_page":f"{items_per_page}","tab":"indicators","region":"","countries":"","cities":"","search":"","star":"","sort_by":"overallscore","order_by":"asc","program_type":""}

    payload = ""
    headers = ""

    # Gửi yêu cầu kết nối
    response = requests.request("GET", url, data=payload, headers=headers, params=querystring)
    
    # Nếu trả về Status 200 --> Success
    if response.status_code == 200:
        text = response.json()
        
        list_uni = text['score_nodes']  

        for i in range(items_per_page):
            # Tên đại học
            name.append(list_uni[i]['title'])

            # Châu lục
            region.append(list_uni[i]['region'])
            
            # Vị trí
            country.append(list_uni[i]['country'])
            
            # Điểm trung bình
            overall.append(list_uni[i]['overall_score'])
            
            # Danh sách điểm chi tiết
            scores = list_uni[i]['scores']
            
            AcademicReputation.append(scores[0]['score'])
            EmployerReputation.append(scores[1]['score'])
            FacultyStudentRatio.append(scores[2]['score'])
            CitationsPerFaculty.append(scores[3]['score'])
            InternationalFacultyRatio.append(scores[4]['score'])
            InternationalStudentsRatio.append(scores[5]['score'])
            InternationalResearchNetwork.append(scores[6]['score'])
            EmploymentOutcomes.append(scores[7]['score'])
            Sustainability.append(scores[8]['score'])
                
    data = pd.DataFrame({
        "Institution Name": name,
        "Region": region,
        "Country": country, 
        "Overall": overall,
        "Academic Reputation": AcademicReputation,
        "Employer Reputation": EmployerReputation,
        "Faculty Student Ratio": FacultyStudentRatio,
        "Citations Per Faculty": CitationsPerFaculty,
        "International Faculty Ratio": InternationalFacultyRatio,
        "International Students Ratio": InternationalStudentsRatio,
        "International Research Network": InternationalResearchNetwork,
        "Employment Outcomes": EmploymentOutcomes,
        "Sustainability": Sustainability,
    })
    
    return data



def Generate_Universities_Dataset(quantity, items_per_page):
    index = 0
    num_uni = 0
    data = pd.DataFrame()
    
    # Nếu số lượng vượt quá yêu cầu thì kết thúc
    while num_uni < quantity:
        data = pd.concat([data, collect_data(index, items_per_page)], ignore_index=True)
        
        num_uni += items_per_page
        index += 1

    return data

# Lưu data vào file csv
uni_df = Generate_Universities_Dataset(1200, 15)

file_path = '../Data/raw/universities.csv'

uni_df.to_csv(file_path, index=False)

print(f'Data saved to {file_path}')

# DataFrame World Universities Ranking
uni_df
