==> data_module/4.0-Modeling.ipynb <==
<div style="text-align: center; background-color: #b1d1ff; font-family: 'Trebuchet MS', Arial, sans-serif; color: white; padding: 20px; font-size: 40px; font-weight: bold; border-radius: 0 0 0 0; box-shadow: 0px 6px 8px rgba(0, 0, 0, 0.2);">
Modeling 
</div>

import pandas as pd 
import numpy as np 
from sklearn.preprocessing import LabelEncoder
from sklearn.model_selection import train_test_split
from sklearn.linear_model import LinearRegression
from sklearn.metrics import r2_score
from sklearn.preprocessing import PolynomialFeatures
from sklearn.ensemble import RandomForestRegressor
from sklearn.neural_network import MLPRegressor
import bentoml

df = pd.read_csv(r"../Data/processed/universities_processed.csv")
df.dtypes
df.info()

# Số hoá dữ liệu kiểu object

labelencoder = LabelEncoder()
df['Region_code'] = labelencoder.fit_transform(df['Region'].values)
df['Country_code'] = labelencoder.fit_transform(df['Country'].values)
y = np.array(df.iloc[0:602,3])
X = np.array(df.iloc[0:602,[4, 5, 6, 7, 8, 9, 10,11, 12]].values)

# Phân tách tập dữ liệu thành tập huấn luyện và tập kiểm thử

X_train, X_test, y_train, y_test = train_test_split(X,y,test_size=0.3,random_state=0)

# Hồi quy tuyến tính

linear_X_train = X_train
linear_y_train = y_train

regressor = LinearRegression()
regressor.fit(linear_X_train,linear_y_train)

linear_y_pred = regressor.predict(X_test)


r2_score(y_test,linear_y_pred)

## Nhận xét:
Độ chính xác quá thấp nên không nhận mô hình này.

# Hồi quy bình phương

square = PolynomialFeatures(degree=2,include_bias=False)
square_X_train = square.fit_transform(X_train)
square_y_train = y_train

square_model = LinearRegression()
square_model.fit(square_X_train,square_y_train)

square_X_test = square.fit_transform(X_test)
square_y_pred  = square_model.predict(square_X_test)

r2_score(y_test,square_y_pred)

# Random Forest Regression

randomforestregressor = RandomForestRegressor(n_estimators=200,criterion='squared_error')
randomforestregressor.fit(X_train,y_train)

random_forest_y_pred = randomforestregressor.predict(X_test)

r2_score(y_test,random_forest_y_pred)

# Mạng Nơ ron


mplregressor = MLPRegressor(hidden_layer_sizes=(100,100),activation='relu',max_iter=300)
mplregressor.fit(X_train,y_train)
nn_y_pred = mplregressor.predict(X_test)
r2_score(y_test,nn_y_pred)

X_miss = np.array(df.iloc[602:1200,[4, 5, 6, 7, 8, 9, 10,11, 12]].values)
y_pred = regressor.predict(X_miss)
for i in range(602,1199):
    df.at[i,'Overall'] = y_pred[i-602]

df.info()
